add tests

+++ src/main/java/com/innowise/orderservice/dto/OrderItemDto.java
@@ -20,5 +20,5 @@ public class OrderItemDto {
 
     @NotNull(message = "Quantity is required")
     @Min(value = 1, message = "Quantity must be at least 1")
-    private Integer quantity;
+    private Short quantity;
 }
